Transfer part of the implementation logic to pinctrl
